requirements.txt
# Data source
alpaca-py==0.26.0

# GCP
google-cloud-storage==2.17.0

# Data processing
pandas==2.2.2
pyarrow==16.0.0

# Config & secrets
python-dotenv==1.0.1

# Code quality
pytest==8.2.0
black==24.4.2
